[modify] add main tokenization
