+++ includes/token.h
@@ -6,7 +6,7 @@
 /*   By: jeongmin <jeongmin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/10 16:01:53 by jeongmin          #+#    #+#             */
-/*   Updated: 2023/01/10 22:03:10 by jeongmin         ###   ########.fr       */
+/*   Updated: 2023/01/11 17:06:41 by jeongmin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,8 @@ typedef struct s_token
 }	t_token;
 
 // token.c
-t_error	tokenization(char *line);
+void	del_t_token(void *content);
+t_list	*tokenization(char *line);
 
 // arr.c
 int		*init_arr(size_t len);
