+++ main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: wonyang <wonyang@student.42seoul.kr>       +#+  +:+       +#+        */
+/*   By: jeongmin <jeongmin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/08 22:30:32 by wonyang           #+#    #+#             */
-/*   Updated: 2023/01/11 11:20:47 by wonyang          ###   ########seoul.kr  */
+/*   Updated: 2023/01/11 17:05:51 by jeongmin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,14 +19,28 @@
 #include "minishell.h"
 #include "token.h"
 
+static void	print_lst(t_list *lst)
+{
+	t_token	*token;
+
+	while (lst)
+	{
+		token = (t_token *)(lst->content);
+		printf("[%d] [%s]\n", token->type, token->str);
+		lst = lst->next;
+	}
+}
+
 int	main(int argc, char **argv, char **env)
 {
 	char	*str;
+	t_list	*lst;
 
 	(void)argc;
 	(void)argv;
 	(void)env;
 	set_minishell_setting();
+	lst = NULL;
 	while (1)
 	{
 		str = readline("prompt$ ");
@@ -36,7 +50,14 @@ int	main(int argc, char **argv, char **env)
 			continue ;
 		add_history(str);
 		printf("%s\n", str);
-		tokenization(str);
+		lst = tokenization(str);
+		if (!lst)
+		{
+			free(str);
+			continue ;
+		}
+		print_lst(lst->next);
+		ft_lstclear(&lst, del_t_token);
 		free(str);
 	}
 	return (0);
